+++ derive/tests/scalar.rs
@@ -1,0 +1,7 @@
+mod schema_utils;
+mod scalar {
+    mod common;
+    mod input_type_tests;
+    mod output_type_tests;
+    mod scalar_tests;
+}